Move authUrls to Squared
This makes these utility functions more accessible as we redesign the
Gafaelfawr user menu to use the PrimaryNav component.
