+++ .changeset/friendly-beans-juggle.md
@@ -1,0 +1,6 @@
+---
+'@lsst-sqre/squared': minor
+'squareone': minor
+---
+
+Moved auth URLs into Squared as a library. The `getLoginUrl` and `getLogout` URL functions compute the full URLs to the RSP's login and logout endpoints and include the `?rd` query strings to return the user to current and home URL respectively.
