+++ packages/squared/src/lib/authUrls.ts
